chore: created a modal ui component
