+++ packages/appkit/src/modal/w3m-modal/styles.ts
@@ -1,23 +1,6 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-  outerBackdrop: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)'
-  },
-  innerBackdropTouchable: {
-    flexGrow: 1
-  },
-  modal: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0
-  },
   card: {
     borderBottomLeftRadius: 0,
     borderBottomRightRadius: 0,
